Created a basic class for towers
Created a class that all towers will either utilize directly or inherit from.

Ice Minki in its current state should just be a basic tower, but instead is a class that inherits from it to practice implementation. Note how export variables set in the base class remain exported in the inheritor, but need to be assigned back to the base class in the ready function.