Fixed unset variables in parent class functions
Originally, when parent functions were called, they would complain that the variables are set to null, despite being set in the child class' export variables. I had fixed this by setting each of the parent's variables as the child's manually, but this felt wrong.

As it turned out, including the function in the child script overrides the parent functions. by removing these function, most notably the process(), everything seems to work now.