Spirit Energy and Shop implementation
Players no longer spend money to place towers and equip items. Instead, the player is given an amount of energy at the start of each wave, and each of the previous actions cost 1 of that resource.

Alternatively, the player can spend their energy towards a shop appearing at the end of each wave. There, the player can buy items for money.
[Broken, items do not appear in inventory despite taking money]