Bugfixes: Reimplemented upgrading towers, Reordered tower options
Forgot to reimplement upgrading towers from the inventory when refactoring code. It crashed when attempting to upgrade due to a missing function. The backpack has that function added back.

Tower options used to get frequently blocked by other towers. They should draw on top of all towers now, regardless of the position of the parent tower.