Added Boats and Bombs
As a demonstration of tower placement capabilities, boats and bombs have been added as tower types.

Boats are functionally identical to regular dart towers, but have larger range and can only be placed in water.

Bombs can only be placed on paths. Their range is only slightly larger than their body, but when they explode, their explosion deals major damage. The bomb is lost after exploding, so the upgrade menu is hidden for this tower.