Big change to Tower placement mechanics
Towers Panels now mostly just emit signals instead of handling the placement process.
Instead, the container is the new panel manager, handling the placement logic.

Additionally, now that placement can be linked back to a certain tower, a placement limit is in effect. All towers are now limited to one copy of that type on the board. Items to expand this limit are planned.