Fixed hotkey double-press requirement
Pressing a hotkey while a tower is held used to just clear the held tower. The 2nd press would then load a tower.
However, clicking a panel with a held tower wouldn't have this delay because the panel sends the clicked signal after the held is cleared.

Removing the Return after the data is cleared appears to fix the hotkey issue without affecting clicking