Fixed camera markers, added tower cleanup function
Buttons appear when the wave start button is pressed, and all previous waves are cleared. They disappear after being clicked on.
Relocated camera boundary markers to be children of the camera so that they are properly moved when the camera moves.
Added a function that removes towers that are beyond the area between the two markers (i.e. not in range of the camera)